Adjust

+++ src/frame.rs
@@ -2,7 +2,7 @@ use super::*;
 
 pub(crate) struct Frame {
   pub(crate) filters: usize,
-  pub(crate) fps: Option<f64>,
+  pub(crate) fps: Option<f32>,
   pub(crate) number: u64,
 }
